requirements.txt
Pillow
yeelight